feat(relay.sh & relay-mpc.sh): don't install docker if `NO_DOCKER_INSTALL` passed

+++ relay.sh
@@ -1,24 +1,27 @@
 #!/bin/bash
 set -e
 
-apt-get remove -y docker docker-engine docker.io containerd runc && \
-apt-get update && \
-apt-get install -y \
-    ca-certificates \
-    curl \
-    gnupg \
-    lsb-release
+# install docker by default
+if [[ -z $NO_DOCKER_INSTALL ]]; then
+    apt-get remove -y docker docker-engine docker.io containerd runc && \
+    apt-get update && \
+    apt-get install -y \
+        ca-certificates \
+        curl \
+        gnupg \
+        lsb-release
 
-mkdir -p /etc/apt/keyrings
+    mkdir -p /etc/apt/keyrings
 
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
 
-echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
-  $(lsb_release -cs) stable" | tee /etc/apt/sources.list.d/docker.list > /dev/null
+    echo \
+      "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
+      $(lsb_release -cs) stable" | tee /etc/apt/sources.list.d/docker.list > /dev/null
 
-apt-get update
-apt-get install -y docker-ce docker-ce-cli containerd.io docker-compose-plugin
+    apt-get update
+    apt-get install -y docker-ce docker-ce-cli containerd.io docker-compose-plugin
+fi
 
 echo "Please enter your private key"
